exploding the SQLTransformer at the DAG level
